# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel OpenApi Demo Documentation",
        "description": "L5 Swagger OpenApi description",
        "contact": {
            "email": "prempawar130@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/create": {
            "post": {
                "tags": [
                    "MongoDb project"
                ],
                "summary": "Create Data in MongoDB",
                "description": "create<br/>\n       Success Code:<br/>\n       10012: Data inserted successfully in MongoDB.<br/>\n       Error Code:<br/>\n       7002: Please enter a valid session-token.<br/>\n       7035: Please enter a valid firstname.<br/>\n       7036: Please enter a valid lastname.<br/>\n       7037: Please enter a valid phone.<br/>\n       ",
                "operationId": "App\\Http\\Controllers\\mongoDbController::crateRegistration",
                "parameters": [
                    {
                        "name": "session-token",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Provide Folder details",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "firstname",
                                    "lastname",
                                    "phone"
                                ],
                                "properties": {
                                    "firstname": {
                                        "type": "string",
                                        "example": "prem"
                                    },
                                    "lastname": {
                                        "type": "string",
                                        "example": "pawar"
                                    },
                                    "Phone": {
                                        "type": "string",
                                        "example": "8407595585"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "422": {
                        "description": "Wrong credentials response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Sorry, wrong email address or password. Please try again"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/read": {
            "get": {
                "tags": [
                    "MongoDb project"
                ],
                "summary": "List all data in mongoDB newdata collection",
                "description": "fetch collection value<br/>\n      Success Code:<br/>\n            10005: Folder list fetch successfull.<br/>\n      Error Code:<br/>\n            7002: Please enter a valid session-token.<br/>\n            7006: Please enter a valid folder_state(A/I).<br/>\n            7007: pleasse enter a valid folder_type(NORMAL/DRIP).<br/> \n       ",
                "operationId": "App\\Http\\Controllers\\mongoDbController::readData",
                "parameters": [
                    {
                        "name": "session-token",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "422": {
                        "description": "Wrong credentials response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Sorry, wrong email address or password. Please try again"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/updateData": {
            "put": {
                "tags": [
                    "MongoDb project"
                ],
                "summary": "Update existing project",
                "description": "update data",
                "operationId": "update data mongodb in collection",
                "parameters": [
                    {
                        "name": "session-token",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "session-token",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Provide data details",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "student_fname",
                                    "student_lname",
                                    "student_phone"
                                ],
                                "properties": {
                                    "mongo_id": {
                                        "type": "string",
                                        "example": "213232131"
                                    },
                                    "firstname": {
                                        "type": "string",
                                        "example": "john"
                                    },
                                    "lastname": {
                                        "type": "string",
                                        "example": "doe"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": "9784"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
        "/api/deleteData": {
            "delete": {
                "tags": [
                    "MongoDb project"
                ],
                "summary": "delete data in mongoDB",
                "description": "Deletes a record and returns no content",
                "operationId": "delete data in mongoDB",
                "parameters": [
                    {
                        "name": "mongo_id",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "session-token",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
...
        "/api/swa_parameter": {
            "get": {
                "tags": [
                    "PREM Tag"
                ],
                "summary": "Get list of available subscription plans for user",
                "description": "Subscription list",
                "operationId": "App\\Http\\Controllers\\parametercontroller::parameter",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "components/schemas/ProjectResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
...
    },
...
    "security": [
        []
    ]
}